api/tempestas_api/urls.py
from django.conf.urls import url, include
from django.contrib import admin
from rest_framework_simplejwt.views import (
    TokenObtainPairView,
    TokenRefreshView,
)

from .views import change_password

urlpatterns = [
    url('', include('wx.urls')),
    url(r'^admin/', admin.site.urls),
    url(r'^accounts/', include('django.contrib.auth.urls')),
    url(r'^api-auth/', include('rest_framework.urls', namespace='rest_framework')),
    url(r'^api/token/$', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    url(r'^api/token/refresh/$', TokenRefreshView.as_view(), name='token_refresh'),
    url(r'^password/$', change_password, name='change_password'),
]

admin.site.site_header = 'Surface Admin Area'
admin.site.site_title = admin.site.site_header
